Get rid of link to minify repo
Along with the buttons, it creates too much clutter.
I'd actually want the github repo link to be right of the h1, but markdown
creates a p around the link, which makes it hard to properly inline them.
Anyway, let's just remove the link for now, it's already at the end of the post
as well.