Stop the spam
I was getting a lot of spam.
Probably just Akismet that stopped working - which I've also reactivated.
But I figured I might submit the form with JS as well; that should
probably help reduce the amount of spam as well. Plus, it lets me
update the form with a success/error message.